Fixes and a new counter
Added a new counter type that might make it easer to name stuff like episodes of a tvshow or something like that.

I also fixed a bug in the recursvie call to the directory iterator that decremented the levels counter when it shouldn't.

The main function should now be easier to understand